RAG-PROJECT/requirements.txt
streamlit
langchain
langchain-groq
chromadb
tiktoken
pypdf
python-docx
python-dotenv
sentence-transformers
